requirements.txt
pandas
matplotlib.pyplot
seaborn
sys
ast
wordcloud